+++ src/test/java/ca/corefacility/bioinformatics/irida/processing/concatenate/SequencingObjectConcatenatorFactoryTest.java
@@ -1,0 +1,59 @@
+package ca.corefacility.bioinformatics.irida.processing.concatenate;
+
+import static org.junit.Assert.assertTrue;
+
+import java.util.Set;
+
+import org.junit.Test;
+
+import com.google.common.collect.Sets;
+
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFilePair;
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SingleEndSequenceFile;
+import ca.corefacility.bioinformatics.irida.processing.concatenate.impl.SequenceFilePairConcatenator;
+import ca.corefacility.bioinformatics.irida.processing.concatenate.impl.SingleEndSequenceFileConcatenator;
+
+/**
+ * Unit test for {@link SequencingObjectConcatenatorFactory}
+ */
+public class SequencingObjectConcatenatorFactoryTest {
+	@Test
+	public void testGetConcatenatorSingle() {
+		SequencingObjectConcatenator<SingleEndSequenceFile> concatenator = SequencingObjectConcatenatorFactory
+				.getConcatenator(SingleEndSequenceFile.class);
+		assertTrue(concatenator instanceof SingleEndSequenceFileConcatenator);
+	}
+
+	@Test
+	public void testGetConcatenatorPair() {
+		SequencingObjectConcatenator<SequenceFilePair> concatenator = SequencingObjectConcatenatorFactory
+				.getConcatenator(SequenceFilePair.class);
+		assertTrue(concatenator instanceof SequenceFilePairConcatenator);
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testGetConcatenatorError() {
+		SequencingObjectConcatenatorFactory.getConcatenator(SequencingObject.class);
+	}
+
+	@Test
+	public void testGetConcatenatorSingleCollection() {
+		Set<SingleEndSequenceFile> fileSet = Sets.newHashSet(new SingleEndSequenceFile(null));
+		SequencingObjectConcatenator<?> concatenator = SequencingObjectConcatenatorFactory.getConcatenator(fileSet);
+		assertTrue(concatenator instanceof SingleEndSequenceFileConcatenator);
+	}
+
+	@Test
+	public void testGetConcatenatorPairCollection() {
+		Set<SequenceFilePair> fileSet = Sets.newHashSet(new SequenceFilePair());
+		SequencingObjectConcatenator<?> concatenator = SequencingObjectConcatenatorFactory.getConcatenator(fileSet);
+		assertTrue(concatenator instanceof SequenceFilePairConcatenator);
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testGetConcatenatorMixedError() {
+		Set<SequencingObject> fileSet = Sets.newHashSet(new SequenceFilePair(), new SingleEndSequenceFile(null));
+		SequencingObjectConcatenatorFactory.getConcatenator(fileSet);
+	}
+}